wip: module restructure
